pyproject.toml
[tool.poetry]
name = "bike-rent"
version = "0.1.0"
description = ""
authors = ["Julia Kuznetsova <kooznitsa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
Jinja2 = "^3.1.4"
pytz = "^2024.1"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"
drf-registration = "^0.1.3"
pytest-django = "^4.8.0"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.6.0"
django-redis = "^5.4.0"
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
